requirements.txt
# DockShield Requirements
# Core dependencies
PyQt6>=6.5.0
docker>=7.0.0
schedule>=1.2.0
paramiko>=3.4.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
psutil>=5.9.0
cryptography>=41.0.0

# Optional cloud storage support
# Uncomment if needed:
# boto3>=1.28.0
# google-cloud-storage>=2.10.0
# azure-storage-blob>=12.19.0
